ai-agents/requirements.txt
# Project Artha - Multi-Agent Financial AI System Requirements
google-adk>=0.1.0
google-genai>=0.3.0
aiohttp>=3.8.0
asyncio>=3.4.3
python-dotenv>=1.0.0
dataclasses>=0.6
typing-extensions>=4.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0

-e .